Improved README.md, added deleting of realm and client

+++ provision/keycloak.py
@@ -3,6 +3,12 @@ import json
 from pprint import pprint as pp
 from dynaconf import Dynaconf
 
+
+class KeycloakAPI:
+    def __init__(self, settings: Dynaconf):
+        pass
+
+
 s = Dynaconf(
     environments=True,
     settings_file='keycloak.yaml'
@@ -11,11 +17,42 @@ s = Dynaconf(
 databricks_base_url = f'https://{s.databricks_deployment}.cloud.databricks.com'
 databricks_saml_url = f'{databricks_base_url}/saml/consume'
 
+keycloak_api = KeycloakAPI(s)
+
+
+def get_realms():
+    return requests.get(
+        f'{s.keycloak_base_url}/auth/admin/realms',
+        headers={
+            'content-type': 'application/json',
+            'Authorization': f'bearer {token}'
+        }
+    ).json()
+
+
+def get_realm():
+    return [realm for realm in get_realms() if s.keycloak_realm == realm['realm']]
+
+
+def get_clients():
+    return requests.get(
+        f'{s.keycloak_base_url}/auth/admin/realms/{s.keycloak_realm}/clients',
+        headers={
+            'content-type': 'application/json',
+            'Authorization': f'bearer {token}'
+        }
+    ).json()
+
+
+def get_client() -> list:
+    return [client for client in get_clients() if s.keycloak_client_id == client['clientId']]
+
 
 def parse_result(action: str, result):
     try:
         if not result.ok:
             raise Exception(f'{action}: {result.text} {result.status_code}')
+        print(f'{action}: {result.text} {result.status_code}')
     except Exception as ex:
         pp(ex)
 
@@ -30,6 +67,16 @@ token = json.loads(requests.post(
         'password': s.keycloak_password
     }
 ).text)['access_token']
+
+# Delete existing realm if present
+if 1 == len(get_realm()):
+    api_result = requests.delete(
+        f'{s.keycloak_base_url}/auth/admin/realms/{s.keycloak_realm}',
+        headers={
+            'Authorization': f'bearer {token}'
+        }
+    )
+    parse_result('Delete realm', api_result)
 
 # Create realm
 api_result = requests.post(
@@ -46,13 +93,14 @@ api_result = requests.post(
 parse_result('Create realm', api_result)
 
 # Delete existing client if present
-api_result = requests.delete(
-    f'{s.keycloak_base_url}/auth/admin/realms/{s.keycloak_realm}/clients/{s.keycloak_client_id}',
-    headers={
-        'Authorization': f'bearer {token}'
-    }
-)
-parse_result('Delete client', api_result)
+if 1 == len(get_client()):
+    api_result = requests.delete(
+        f'{s.keycloak_base_url}/auth/admin/realms/{s.keycloak_realm}/clients/{s.keycloak_client_id}',
+        headers={
+            'Authorization': f'bearer {token}'
+        }
+    )
+    parse_result('Delete client', api_result)
 
 # Create client
 api_result = requests.post(
